backend/requirements.txt
Django>=5.0.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
whitenoise
gunicorn
dj-database-url
python-dotenv
pandas
numpy
scikit-learn
psycopg2-binary
